refactored program and parsed_program
