+++ src/parser/mod.rs
@@ -1,9 +1,10 @@
 pub mod program;
+mod parsed_program;
 use std::fmt::Display;
 
 #[derive(Debug)]
 pub enum ParseError {
-    FailedToOpenFile(std::io::Error),
+    OpenError(std::io::Error),
     InvalidYaml(serde_yaml::Error),
     InvalidUmask(String, String),
     InvalidSignal(String, String),
@@ -18,7 +19,7 @@ impl From<serde_yaml::Error> for ParseError {
 impl Display for ParseError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            ParseError::FailedToOpenFile(err) => write!(f, "Error opening taskmaster config file: {err}\nConsider making a reload request after creating one"),
+            ParseError::OpenError(err) => write!(f, "Error opening taskmaster config file: {err}\nConsider making a reload request after creating one"),
             ParseError::InvalidYaml(err) => write!(f, "Error parsing taskmaster config file: {err}\nConsider making a reload request after fixing the issue"),
             ParseError::InvalidSignal(sig, prog_name) => write!(f, "Error parsing taskmaster config file: invalid stopsignal {sig} for program {prog_name}\nConsider making a reload request after fixing the issue"),
             ParseError::InvalidUmask(sig, prog_name) => write!(f, "Error parsing taskmaster config file: {sig} for program {prog_name}\nConsider making a reload request after fixing the issue"),
